Start a big cleanup, WIP

+++ source/backend/engine/CarlaEngineJack.cpp
@@ -1031,7 +1031,7 @@ public:
             startThread();
             callback(ENGINE_CALLBACK_ENGINE_STARTED, 0,
                      opts.processMode, opts.transportMode,
-                     pData->sampleRate,
+                     static_cast<float>(pData->sampleRate),
                      getCurrentDriverName());
             return true;
         }
@@ -1825,7 +1825,7 @@ protected:
                             mdataTmp[0] = static_cast<uint8_t>(midiEvent.data[0] | (engineEvent.channel & MIDI_CHANNEL_BIT));
 
                             // copy rest
-                            carla_copy<uint8_t>(mdataTmp+1, midiEvent.data+1, size-1);
+                            carla_copy<uint8_t>(mdataTmp+1, midiEvent.data+1, size-1U);
 
                             // done
                             mdataPtr = mdataTmp;
